requirements.txt
openai>=1.37.1, <2.0.0
python-dotenv>=1.0.1, <2.0.0
ipykernel<=6.8.0
psycopg2-binary
faker
google-cloud-bigquery
azure-search-documents==11.4.0
azure-core
streamlit==1.33.0
flask>=0.12.3
Flask-Cors==3.0.9
Flask-Security==3.0.0
flask-jwt-simple[asymmetric_crypto]
flask[async]
sqlalchemy
pandas
chardet
speechrecognition
pyodbc
